Login and Signup Pages
Signup page works (still need to work on strong password, and suggestions and all of that, but the basic essence of creating a new user and a little bit of validation works)

Login page doesn't. Simple changes required, will be done pretty soon.

Currently validation is being done with php on server side,  but need to change that to JS on browser to make server load less.